create useChallenges hook
